// appth/src/app/models/LlegadaTardeBusqueda.ts
export class LlegadaTardeBusqueda {
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    tipoDocumento: string;
    nroDocumento: string;
    cargo: string;
    area: string;
    contrato: number;
    fecha: Date;
    horaentrada: string;
    detalle: string;
    tipotardanza: number;

}